+++ chat-widget/src/common/telemetry/TelemetryConstants.ts
@@ -119,6 +119,7 @@ export enum TelemetryEvent {
     StartChatSDKCall = "StartChatCall",
     StartChatEventRecevied = "StartChatEventReceived",
     EndChatSDKCall = "EndChatSDKCall",
+    PrepareEndChat = "PrepareEndChat",
     EndChatEventReceived = "EndChatEventReceived",
     WindowClosed = "WindowClosed",
     OnNewMessageFailed = "OnNewMessageFailed",
@@ -129,10 +130,10 @@ export enum TelemetryEvent {
     ErrorUIPaneLoaded = "ErrorUIPaneLoaded",
     DownloadTranscriptFailed = "DownloadTranscriptFailed",
     StartChatFailed = "StartChatFailed",
-    IC3ThreadUpdateEventReceived = "IC3ThreadUpdateEventReceived",
     ConfirmationCancelButtonClicked = "ConfirmationCancelButtonClicked",
     ConfirmationConfirmButtonClicked = "ConfirmationConfirmButtonClicked",
     LoadingPaneLoaded = "LoadingPaneLoaded",
+    StartChatErrorPaneLoaded = "StartChatErrorPaneLoaded",
     EmailTranscriptLoaded = "EmailTranscriptLoaded",
     OutOfOfficePaneLoaded = "OutOfOfficePaneLoaded",
     ConfirmationPaneLoaded = "ConfirmationPaneLoaded",
@@ -267,7 +268,6 @@ export class TelemetryConstants {
             case TelemetryEvent.EmailTranscriptSent:
             case TelemetryEvent.EmailTranscriptFailed:
             case TelemetryEvent.DownloadTranscriptFailed:
-            case TelemetryEvent.IC3ThreadUpdateEventReceived:
             case TelemetryEvent.ConfirmationCancelButtonClicked:
             case TelemetryEvent.ConfirmationConfirmButtonClicked:
             case TelemetryEvent.PreChatSurveyStartChatMethodFailed:
@@ -309,6 +309,7 @@ export class TelemetryConstants {
             case TelemetryEvent.PostChatContextCallFailed:
             case TelemetryEvent.PostChatContextCallSucceed:
             case TelemetryEvent.GetConversationDetailsException:
+            case TelemetryEvent.PrepareEndChat:
                 return ScenarioType.SDK;
 
             case TelemetryEvent.VideoCallAcceptButtonClick: