+++ chat-widget/src/common/utils.ts
@@ -1,4 +1,4 @@
-import { AriaTelemetryConstants, ChatSDKError, Constants, HtmlAttributeNames, LocaleConstants } from "./Constants";
+import { AriaTelemetryConstants, Constants, HtmlAttributeNames, LocaleConstants } from "./Constants";
 import { BroadcastEvent, LogLevel, TelemetryEvent } from "./telemetry/TelemetryConstants";
 
 import { BroadcastService } from "@microsoft/omnichannel-chat-components";
@@ -8,6 +8,7 @@ import { ITimer } from "./interfaces/ITimer";
 import { KeyCodes } from "./KeyCodes";
 import { Md5 } from "md5-typescript";
 import { TelemetryHelper } from "./telemetry/TelemetryHelper";
+import { ChatSDKErrorName } from "@microsoft/omnichannel-chat-sdk";
 
 const getElementBySelector = (selector: string | HTMLElement) => {
     let element: HTMLElement;
@@ -398,7 +399,7 @@ export const getConversationDetailsCall = async (chatSDK: any) => {
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const checkContactIdError = (e: any) => {
-    if (e?.message === ChatSDKError.AuthContactIdNotFoundFailure) {
+    if (e?.message === ChatSDKErrorName.AuthContactIdNotFoundFailure) {
         const contactIdNotFoundErrorEvent: ICustomEvent = {
             eventName: BroadcastEvent.ContactIdNotFound,
             payload: {
@@ -423,4 +424,20 @@ export const createFileAndDownload = (fileName: string, blobData: string, mimeTy
     document.body.appendChild(aElement);
     aElement.click();
     document.body.removeChild(aElement);
+};
+
+/**
+ * 
+ * Replace placeholders with format {0}..{n} , in a string with values
+ * 
+ * @param template String with placeholders to be replaced
+ * @param values array of values to replace the placeholders
+ * @returns formatted string with replaced values
+ */
+// use of any for values as array of any type is passed
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const  formatTemplateString = (templateMessage : string, values : any) => {
+    return templateMessage.replace(/{(\d+)}/g, (match, index) => {
+        return typeof values[index] !== "undefined" ? values[index] : match;
+    });
 };