+++ CHANGE_LOG.md
@@ -6,6 +6,45 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- Enabled customizations for the start chat error pane by adding new `IStartChatErrorPaneProps` interface
+- Added `PrepareEndChat` Telemetry Event to identify end chat workflow
+- Add specific error strings to start chat error pane for authentication related failure scenarios by including new texts to `IStartChatErrorPaneControlProps`
+
+### Fixed
+
+- Fixed an issue, where after the agent end the chat and C2 sees the disconnect banner after toggling, refreshing the browser does not show the message box.
+- Clear ChatSDK's internal `liveChatContext` when `conversationState` is set to `Closed` on `startChat()`
+- A11Y fix for JAWS screen reader, not able to read/navigate messages when scan mode is activated
+- Fix for InitiateEndChat event, which now updates state for confirmation pane, to enable close of the widget when clicking on the close button.
+
+### Changed
+
+- Updated exception details on chat start to include http status code for all `WidgetLoadFailed` and `WidgetLoadComplete` events with error
+- Uptake [@microsoft/omnichannel-chat-components@1.1.0](https://www.npmjs.com/package/@microsoft/omnichannel-chat-components/v/1.1.0)
+- Uptake [@microsoft/omnichannel-chat-sdk@1.6.2](https://www.npmjs.com/package/@microsoft/omnichannel-chat-sdk/v/1.6.2)
+
+## [1.5.0] - 2023-11-21
+
+### Added
+
+- [A11Y] Notification banner when e-mail address is introduced to receive transcript after chat ends
+- Added `skipBroadcastChannelInit` prop to avoid duplicate initialization on BroadcastChannel
+
+### Fixed
+
+- Fixed an issue where invoking `StartChat` event will create different chats across all tabs
+- Fixed an issue where post chat does not load on first chat on a browser session
+- Fixed widget crashing due to `BroadcastChannel` not supported on `Safari` browsers below `15.4`
+- Fixed an issue where Chat disconnect banner message is not showing localized text
+- Fixed an issue where invoking `StartChat` opens pre-chat survey during out of operation hours
+- Fixed an issue where new chat starts even after reconnectUrl is being used
+
+### Changed
+
+- Uptake [@microsoft/omnichannel-chat-components@1.0.8](https://www.npmjs.com/package/@microsoft/omnichannel-chat-components/v/1.0.8)
+- Uptake [@microsoft/omnichannel-chat-sdk@1.5.7](https://www.npmjs.com/package/@microsoft/omnichannel-chat-sdk/v/1.5.7)
+
 ## [1.4.0] - 2023-10-25
 
 ### Changed
@@ -204,6 +243,26 @@ All notable changes to this project will be documented in this file.
 # Chat Components
 
 ## [Unreleased]
+
+## [1.1.0] 2023-12-12
+### Added
+
+- Add ability to customize `PreChatSurveyPane`'s toggle inputs using `IPreChatSurveyPaneToggleInputStyles` & Use `baseline` by default
+
+### Fixed
+
+- Fixed loading pane icon image fit
+
+## [1.0.9] 2023-10-31
+
+### Changed
+- Modify texts for email input validation
+
+## [1.0.8] - 2023-10-31
+
+### Added
+
+- Added tooltips for header and footer action bar buttons
 
 ## [1.0.7] - 2023-9-13